refactor: login_view, login_page 분리
Observe 패턴을 위한 view, page 분리
